Airline Integration/requirements.txt
selenium==4.15.2
robotframework==6.1.1
robotframework-seleniumlibrary==6.2.0
robotframework-retryfailed==0.2.0
robotframework-pabot==2.16.0
openpyxl==3.1.2
pandas~=2.2.1
matplotlib~=3.8.2
Jinja2==3.1.3
xmltodict==0.13.0
#pyautogui==0.9.54
robotframework-excellib==2.0.1
rpaframework-excel==0.1.0
rpaframework-core==11.3.2
robotframework-datadriver==1.10.0
pytz~=2023.3.post1
beautifulsoup4~=4.12.3
lxml~=5.2.1
